Users points fix

+++ Web/Lab4/src/main/java/org/acrighthere/lab4/backend/model/Point.java
@@ -19,7 +19,7 @@ public class Point {
 
     private boolean hit;
 
-    private long executionTime;
+    private double executionTime;
     private LocalDateTime createdAt;
 
     @ManyToOne(fetch = FetchType.LAZY)
